requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
flask-login==0.4.1
requests==2.22.0
bcrypt==3.1.7
pip==19.3.1
pytest==5.2.2
pytest-flask==0.15.0
pytest-mock==1.11.2
werkzeug==0.16.0